move xtrain -> train
