+++ xallennlp/commands/__init__.py
@@ -1,3 +1,3 @@
+import xallennlp.commands.train  # noqa: F401
 import xallennlp.commands.train_with_mlflow  # noqa: F401
 import xallennlp.commands.tune_with_mlflow  # noqa: F401
-import xallennlp.commands.xtrain  # noqa: F401
